% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmbars-package.R
\docType{package}
\name{emmbars-package}
\alias{emmbars}
\alias{emmbars-package}
\title{emmbars: Add Emmeans Significance Brackets to ggplot Objects}
\description{
Provides add_emmeans_pbars(), a convenience wrapper that adds statistical significance brackets to ggplot objects using post hoc comparisons generated by emmeans::pairs() or emmeans::contrast(). Bracket positions are inferred automatically from the plot structure, with support for facets, nested facets, dodged geoms, and fill/color aesthetics.
}
\author{
\strong{Maintainer}: Nicholas Edenhoffer \email{NicholasEdenhoffer@users.noreply.github.com}

}
\keyword{internal}
